separate pfnc and sfnc from gendc_separator
